+++ Changelog.md
@@ -28,6 +28,8 @@
 - Added the payment dependent shipping fee calculator
 - Added the `units_sold` and the `last_sale_at` attributes to the master product model (SUM/MAX from variants)
 - Added the `Stockable` interface (Contracts)
+- Added the `Stockable` interface to the `Product` and `MasterProductVariant` models
+- Added the `backorder` field to products and product variants
 - Fixed possible null return type on Billpayer::getName() when is_organization is true but the company name is null
 
 ## 3.x Series